Separate styled components.
