+++ src/components/consistent-nav.tsx
@@ -1,12 +1,6 @@
 import React from 'react';
-import styled from '@emotion/styled';
-import { UnifiedNav } from '@mdb/consistent-nav';
 
-import { layer } from './theme';
-
-const UnifiedNavCustom = styled(UnifiedNav)`
-    z-index: ${layer.superFront};
-`;
+import { UnifiedNavCustom } from '../styled/consistent-nav';
 
 const ConsistentNav: React.FunctionComponent = () => (
     <UnifiedNavCustom position="sticky" floraTheme="default" />
